requirements.txt
# Core Requirements (Always needed)
requests>=2.31.0
python-dotenv>=1.0.0

# Discord Bot Requirements
discord.py>=2.3.0
aiohttp>=3.8.0